requirements.txt
# Base dependencies
flask==2.0.1
flask-cors==3.0.10
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
python-dotenv>=0.19.0
requests>=2.23.0

# Model dependencies
ultralytics>=8.0.0  # YOLOv8
torch>=1.7.0
torchvision>=0.8.1

# Nutrition data processing
pandas>=1.1.4
